+++ R/lift.R
@@ -219,4 +219,4 @@ lift = function(input = NULL, output_dir = NULL) {
     }
   }
 
-}
+  }