+++ R/utils.R
@@ -11,7 +11,7 @@ file_name = function (x) basename(normalizePath(x))
 sans_ext = tools::file_path_sans_ext
 file_name_sans = function (x) basename(sans_ext(normalizePath(x)))
 
-# generate uuid for Docker container names
+# generate UUID for Docker container names
 # derived from http://stackoverflow.com/questions/10492817/
 uuid = function () {
   id = paste(sample(c(letters[1:6], 0:9), 30, replace = TRUE),